+++ src/main/kotlin/org/jetbrains/intellij/dependency/CustomPluginsRepository.kt
@@ -7,45 +7,35 @@ import com.jetbrains.plugin.structure.intellij.repository.CustomPluginRepository
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.create
 import org.jetbrains.intellij.debug
+import org.jetbrains.intellij.resolveRedirection
 import org.jetbrains.intellij.utils.DependenciesDownloader
 import org.jetbrains.intellij.utils.ivyRepository
 import java.io.File
-import java.net.URI
 import java.net.URL
 
-class CustomPluginsRepository(repositoryUrl: String, private val dependenciesDownloader: DependenciesDownloader) : PluginsRepository {
-
-    private val pluginsXmlUri: URI
-    private val repositoryUrl: String
-
-    init {
-        val uri = URI(repositoryUrl)
-        if (uri.path.endsWith(".xml")) {
-            this.repositoryUrl = repositoryUrl.substring(0, repositoryUrl.lastIndexOf('/'))
-            pluginsXmlUri = uri
-        } else {
-            this.repositoryUrl = repositoryUrl
-            pluginsXmlUri = uri.run { URI(scheme, userInfo, host, port, "$path/", query, fragment).resolve("updatePlugins.xml") }
-        }
-    }
+class CustomPluginsRepository(
+    private val repositoryUrl: String,
+    private val dependenciesDownloader: DependenciesDownloader,
+) : PluginsRepository {
 
     override fun resolve(project: Project, plugin: PluginDependencyNotation, context: String?): File? {
-        debug(context, "Loading list of plugins from: $pluginsXmlUri")
-        val url = pluginsXmlUri.toURL()
+        debug(context, "Loading list of plugins from: $repositoryUrl")
+        val url = URL(repositoryUrl)
         val downloadUrl = resolveDownloadUrl(url, plugin, CustomPluginRepositoryListingType.PLUGIN_REPOSITORY)
             ?: resolveDownloadUrl(url, plugin, CustomPluginRepositoryListingType.SIMPLE)
             ?: return null
 
-        return downloadZipArtifact(project, downloadUrl, plugin, context)
+        return downloadZipArtifact(downloadUrl, plugin, context)
     }
 
     private fun resolveDownloadUrl(url: URL, plugin: PluginDependencyNotation, type: CustomPluginRepositoryListingType) =
         CustomPluginRepositoryListingParser
-            .parseListOfPlugins(url.readText(), url, URL(repositoryUrl), type)
+            .parseListOfPlugins(url.readText(), url, url, type)
             .find { it.pluginId.equals(plugin.id, true) && it.version.equals(plugin.version, true) }
             ?.downloadUrl
+            ?.resolveRedirection()
 
-    private fun downloadZipArtifact(project: Project, url: URL, plugin: PluginDependencyNotation, context: String?) =
+    private fun downloadZipArtifact(url: URL, plugin: PluginDependencyNotation, context: String?) =
         dependenciesDownloader.downloadFromRepository(context, {
             create(
                 group = "com.jetbrains.plugins",
@@ -54,7 +44,7 @@ class CustomPluginsRepository(repositoryUrl: String, private val dependenciesDow
                 ext = "zip",
             )
         }, {
-            ivyRepository(url.toString())
+            ivyRepository(url)
         }).first()
 
     override fun postResolve(project: Project, context: String?) {