+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## [next]
 
-### Fixed
+### Changed
+- `custom(String)` helper of `intellij.pluginRepositories` configuration requires now passing a direct URL to the `updatePlugins.xml` file [#1252](../../issues/1252)
 
+### Fixed
 - Replace `Contents/Contents` part within JVM arguments to a single `Contents` – happens with macOS distribution
 - `--offline` prevents from using JBR even if it is already downloaded [#1251](../../issues/1251)